Moved Search to components and styled dropdown
