+++ client/src/components/Search.js
@@ -1,4 +1,4 @@
-import './header.scss';
+import './search/search.scss';
 
 import React, { useState, useEffect } from 'react';
 import { Icon } from 'semantic-ui-react';
@@ -8,7 +8,7 @@ import _ from 'lodash';
 import { Subject } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
 
-import SearchDropdown from './SearchDropdown';
+import SearchDropdown from './search/SearchDropdown';
 
 class Search extends React.Component {
   constructor(props) {
@@ -56,7 +56,7 @@ class Search extends React.Component {
   render() {
     const { isSearching, searchResults, text } = this.state;
     return (
-      <div className="search-field-container">
+      <div className="search-form-container">
         <Icon
           name="search"
           size="large"
@@ -64,7 +64,7 @@ class Search extends React.Component {
           onClick={this.clickToShowSearchInput}
         />
         {isSearching && (
-          <div className="search-form-container">
+          <div className="search-field-container">
             <form>
               <input
                 type="text"
@@ -82,6 +82,77 @@ class Search extends React.Component {
     );
   }
 }
+
+// function Search(props) {
+//   const [isSearching, setSearchInput] = useState(false);
+//   const [text, setText] = useState('');
+//   const [searchResults, setSearchResults] = useState([]);
+//   const input = new Subject();
+
+//   const clickToShowSearchInput = () => {
+//     if (!isSearching) {
+//       setSearchInput(true);
+//     }
+//   };
+
+//   const _executeSearch = async () => {
+//     try {
+//       const result = await props.client.query({
+//         query: searchQuery,
+//         variables: { text },
+//         fetchPolicy: 'no-cache',
+//       });
+//       const {
+//         data: { search },
+//       } = result;
+//       console.log(search);
+//       setSearchResults(search);
+//     } catch (error) {
+//       console.log(error);
+//     }
+//   };
+
+//   const onChange = event => {
+//     const { value } = event.target;
+//     if (value !== text) {
+//       input.next();
+//     }
+//     setText(value);
+//   };
+
+//   useEffect(() => {
+//     input.pipe(debounceTime(3000)).subscribe(_executeSearch);
+//     return () => {
+//       input.unsubscribe();
+//     };
+//   });
+
+//   return (
+//     <div className="search-field-container">
+//       <Icon
+//         name="search"
+//         size="large"
+//         color="grey"
+//         onClick={clickToShowSearchInput}
+//       />
+//       {isSearching && (
+//         <div className="search-form-container">
+//           <form>
+//             <input
+//               type="text"
+//               placeholder="Search Principlez"
+//               // onChange={e => setText(e.target.value)}
+//               onChange={onChange}
+//             />
+//           </form>
+//           {/* {!_.isEmpty(searchResults) && (
+//             <SearchResultsDropdown results={searchResults} />
+//           )} */}
+//         </div>
+//       )}
+//     </div>
+//   );
+// }
 
 const searchQuery = gql`
   query searchQuery($text: String!) {
