--- client/src/components/search/StorySearchResult.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function StorySearchResult(props) {
-  return <div>hello stories</div>;
-}
