+++ client/src/components/search/search.scss
@@ -1,0 +1,42 @@
+.search-form-container {
+  margin-left: auto;
+  padding: 20px;
+
+  input {
+    background-color: transparent;
+    border-radius: 10px;
+    padding: 8px;
+    border: 1px solid grey;
+    color: white;
+  }
+
+  input::placeholder {
+    color: white;
+  }
+
+  .search-field-container {
+    display: inline-block;
+  }
+}
+
+.search-dropdown-container {
+  top: 70px;
+  position: absolute;
+  list-style: none;
+  background-color: white;
+  border: 1px solid #f0f0f0;
+  border-radius: 5px;
+  width: 20%;
+  text-align: left;
+  padding: 12px;
+  color: grey;
+
+  h4 {
+    padding: 5px;
+  }
+
+  li {
+    width: 100%;
+    font-size: 18px;
+  }
+}
