+++ client/src/containers/header/index.js
@@ -1,4 +1,3 @@
-export Search from './Search';
 export User from './User';
 export UserProfile from './UserProfile';
 export HeaderDropdown from './HeaderDropdown';
